+++ namer/dnssrv/src/test/scala/io/buoyant/namer/dnssrv/DnsSrvNamerTest.scala
@@ -1,14 +1,11 @@
 package io.buoyant.namer.dnssrv
 
 import com.twitter.finagle._
-import com.twitter.finagle.stats.NullStatsReceiver
 import com.twitter.finagle.util.LoadService
-import com.twitter.util.{Duration, NullTimer}
 import io.buoyant.config.Parser
 import io.buoyant.namer.{NamerConfig, NamerInitializer}
 import io.buoyant.test.FunSuite
 import org.scalatest.Matchers
-import org.xbill.DNS
 
 class DnsSrvNamerTest extends FunSuite with Matchers {
 
@@ -34,20 +31,8 @@ class DnsSrvNamerTest extends FunSuite with Matchers {
       """.stripMargin
 
     val mapper = Parser.objectMapper(yaml, Iterable(Seq(DnsSrvNamerInitializer)))
-    val curator = mapper.readValue[NamerConfig](yaml).asInstanceOf[DnsSrvNamerConfig]
-    assert(curator.refreshIntervalSeconds === Some(60))
-    assert(curator.dnsHosts === Some(Seq("localhost")))
-  }
-
-  test("can resolve some public SRV revord") {
-    val namer = new DnsSrvNamer(Path.empty, new DNS.ExtendedResolver, new NullTimer, Duration.Zero, new NullStatsReceiver)
-    val result = namer.lookupSrv("_http._tcp.mxtoolbox.com.", Path.read("/foo"), null)
-    result.get().simplified match {
-      case NameTree.Leaf(Name.Bound(varAddr)) => varAddr.sample() match {
-        case Addr.Bound(addrs, _) => addrs should not be empty
-        case a => fail(s"unexpected addr: $a")
-      }
-      case x => fail(s"unexpected result: $x")
-    }
+    val config = mapper.readValue[NamerConfig](yaml).asInstanceOf[DnsSrvNamerConfig]
+    assert(config.refreshIntervalSeconds === Some(60))
+    assert(config.dnsHosts === Some(Seq("localhost")))
   }
 }
